# openapi_server/models/response_upload_text_model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class ResponseUploadTextModel(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, code=None, message=None):  # noqa: E501
        """ResponseUploadTextModel - a model defined in OpenAPI

        :param code: The code of this ResponseUploadTextModel.  # noqa: E501
        :type code: int
        :param message: The message of this ResponseUploadTextModel.  # noqa: E501
        :type message: str
        """
        self.openapi_types = {
            'code': int,
            'message': str
        }

        self.attribute_map = {
            'code': 'code',
            'message': 'message'
        }

        self._code = code
        self._message = message

    @classmethod
    def from_dict(cls, dikt) -> 'ResponseUploadTextModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ResponseUploadTextModel of this ResponseUploadTextModel.  # noqa: E501
        :rtype: ResponseUploadTextModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self):
        """Gets the code of this ResponseUploadTextModel.


        :return: The code of this ResponseUploadTextModel.
        :rtype: int
        """
        return self._code

    @code.setter
    def code(self, code):
        """Sets the code of this ResponseUploadTextModel.


        :param code: The code of this ResponseUploadTextModel.
        :type code: int
        """

        self._code = code

    @property
    def message(self):
        """Gets the message of this ResponseUploadTextModel.


        :return: The message of this ResponseUploadTextModel.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message):
        """Sets the message of this ResponseUploadTextModel.


        :param message: The message of this ResponseUploadTextModel.
        :type message: str
        """

        self._message = message