# openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  description: This is `bsvcontentserver`.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: OpenAPI BsvContent
  version: 1.0.0
servers:
- url: https://bsvcontent.herokuapp.com/v1
...
paths:
...
  /api/upload_text:
    post:
      description: upload text data on Bitcoin SV.
      operationId: api_uploadtext
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUploadTextModel'
        description: upload text data on Bitcoin SV.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponseUploadTextModel'
                type: array
          description: successful operation
        "400":
          content: {}
          description: Invalid status value
      security:
      - api_key: []
      summary: upload text data on Bitcoin SV.
      tags:
      - api
      x-codegen-request-body-name: body
      x-openapi-router-controller: openapi_server.controllers.api_uploadtext_controller
components:
...
  schemas:
    RequestUploadTextModel:
      description: request /api/upload_text
      example:
        mnemonic_words: ""
        message: upload text
      properties:
        mnemonic_words:
          type: string
        message:
          type: string
      title: request /api/upload_text
      type: object
...
    ResponseUploadTextModel:
      description: response /api/add_address
      example:
        code: 0
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      title: response /api/add_address
      type: object
...
  securitySchemes:
    api_key:
      in: header
      name: x-api-key
      type: apiKey
      x-apikeyInfoFunc: openapi_server.controllers.security_controller_.info_from_api_key